requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-appconf==1.0.3
django-compressor==2.4
django-markdownx==3.0.1
django-sass-processor==0.8
gunicorn==19.9.0
idna==2.8
libsass==0.19.4
Markdown==3.2
Pillow==7.0.0
psycopg2==2.8.3
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
